Add files via upload
The main application for the project. You supply longitude and latitude for an area, a buffer for the area to cover, the satellite to use, and the time span. It then uses Google Earth Engine to produce a video which is then input into the model which I trained. The model then identifies the waterbodies in the video and the output video contains the masks for the waterbodies.